Add Abort, and remove Loggedstate for FlightState
Added the start of the Abort command, and cleaned up the Recovery
command.

Due to the redundency of a LoggedState, and the needless passing of a
bunch of parameters, it's a lot more maintainable to just have a
FlightState struct and throw that around, or in this case keep it
global.